refactor: decompose log explainer modules
