+++ src/logExplainer/error.ts
@@ -1,0 +1,8 @@
+export function buildLogExplainerError(
+  status: number,
+  message: string
+): Error & { status: number } {
+  const err = new Error(message) as Error & { status: number }
+  err.status = status
+  return err
+}
